AI chat/langchain-openrouter/requirements.txt
langchain-core>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
openai>=1.0.0
python-dotenv>=1.0.0